prontos para acção8

+++ pdv-backend/src/main/java/com/stechengenharia/pdv_backend/sync/dto/SyncOperacaoDTO.java
@@ -13,7 +13,7 @@ public class SyncOperacaoDTO {
     private String entidade;   // cliente | marca | categoria | servico | produto | pedido
     private String operacao;   // CREATE | UPDATE | DELETE
     private String localId;    // UUID gerado no Flutter (CREATE); null em UPDATE/DELETE
-    private Integer id;        // ID real do backend (UPDATE/DELETE)
+     private Long id;       // ID real do backend (UPDATE/DELETE)
 
     /** Payload dinâmico — campos variam por entidade */
     private Map<String, Object> payload = new LinkedHashMap<>();
